fixed some addtolog calls in the early of the file. So they now report as not having found something if nothing was found. instead of just like listing the empty list.
I want to add a boolean or something there so the result of each step can be presented and accessed easily without having to look through it. So the log can stay empty if something passes. not sure whats better.

I want to add a way to verify individual lines. So people can check it before adding to the spreadsheet. Shouldn't be too hard. once everything is a bit more defined.

I want to split up analyze_csv into a few methods. or combine it with some other methods. but I think splitting it up and passing dataframes around is the best way. So each step can be customized as needed