big commit! There is now stuff to detect misreported games. It does this by comparing the actual winners of games as reported by the importer against the winner reported in the rounds file. This was a little tricky because the names in the rounds file are like discord names. and the ones in the importer are the names used in the battles on showdown which might be slightly to completely different. I got around this by instituiting a json that handles aliases between names.
It is clunky and will be trimmed down. but right now it prints the mismatched bad games to reduced_mismatch.csv

This is looking good. I think this is everything that has been asked.
I need to add an allowlist for the mismatch stuff. some games look wrong but a player was disqualified. and sometimes theres other weird stuff that need exceptions