cleaned up a little bit more. got rid of unneeded methods and stuff. as well as cleaning up some scrap code I used to merge some files and stuff.
Code is looking great.

Next step is fix the stuff in the main method by making it its own methods. that add to the log. Then back to TODO list.

Will need to be ready for the bracket which will likely have a different set up entirely which will require different xcl file manipulation.